// agentski-frontend/src/app/models/Pricing.ts
//import { User } from './User';
//import { Company } from './Company';
    export class Pricing {
        distanceLimit: number;
        regularPrice: number;
        overusePrice: number;
        collisionDamage: number;
        discountDays: number;
        discountPercent: number;
        name: string;
        //owner: User;
        //company: Company;
        deleted: boolean;
    }